# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiMessage {
  error: string;
}

export interface DojoOverview {
  totalLessons: number;
  startedLessons: number;
  totalCommits: number;
  /** Whether the git-dojo folder exists at the expected location */
  dojoFound: boolean;
}

export interface Lesson {
  /** Short lesson id such as lesson-01 */
  id: string;
  number: number;
  title: string;
  /** Lesson folder name such as lesson-01-first-snapshot */
  folderName: string;
  /** setup.sh has been run for this lesson */
  hasPlayground: boolean;
  /** The playground contains a git repository */
  initialized: boolean;
  commitCount: number;
  /** The lesson has a scripted teammate (bot.sh) that can act when time passes */
  hasBot: boolean;
}

export type RepoFileStatus = typeof RepoFileStatus[keyof typeof RepoFileStatus];


export const RepoFileStatus = {
  staged: 'staged',
  modified: 'modified',
  staged_and_modified: 'staged_and_modified',
  untracked: 'untracked',
  deleted: 'deleted',
  conflicted: 'conflicted',
} as const;

export interface RepoFile {
  path: string;
  status: RepoFileStatus;
}

export interface RepoCommit {
  hash: string;
  shortHash: string;
  subject: string;
  authorName: string;
  /** ISO 8601 author date */
  date: string;
  /** Branch or tag names pointing at this commit */
  refs: string[];
  /**
     * Branch this commit was listed under, when walking per-branch history
     * @nullable
     */
  branch?: string | null;
}

export interface RepoBranch {
  name: string;
  isCurrent: boolean;
  headHash: string;
}

export interface RepoState {
  lessonId: string;
  hasPlayground: boolean;
  initialized: boolean;
  /** @nullable */
  currentBranch: string | null;
  detachedHead: boolean;
  mergeInProgress: boolean;
  files: RepoFile[];
  commits: RepoCommit[];
  branches: RepoBranch[];
  remotes: string[];
  /** Plain-English explanation of the current repository state */
  summary: string;
  /** The lesson has a scripted teammate (bot.sh) that can act when time passes */
  hasBot: boolean;
}

export type ProgressEntryTrack = typeof ProgressEntryTrack[keyof typeof ProgressEntryTrack];


export const ProgressEntryTrack = {
  visual: 'visual',
  cli: 'cli',
  live: 'live',
} as const;

export interface ProgressEntry {
  /** Module id such as 1.1 (Track A) or lesson-01 (Track B) */
  moduleId: string;
  track: ProgressEntryTrack;
  /** ISO 8601 timestamp of when the module first passed */
  completedAt: string;
}

export interface Progress {
  entries: ProgressEntry[];
}

export type CompleteModuleRequestTrack = typeof CompleteModuleRequestTrack[keyof typeof CompleteModuleRequestTrack];


export const CompleteModuleRequestTrack = {
  visual: 'visual',
  cli: 'cli',
  live: 'live',
} as const;

export interface CompleteModuleRequest {
  moduleId: string;
  track: CompleteModuleRequestTrack;
}

export interface BotActionResult {
  /** Whether the teammate script ran */
  ran: boolean;
  /** What the teammate did, in plain English */
  output: string;
}

export interface CapstoneRepo {
  name: string;
  /** owner/name */
  fullName: string;
  htmlUrl: string;
  cloneUrl: string;
  defaultBranch: string;
}

export interface CapstoneMission {
  /** Mission id (push-commit, create-branch, merge-pr) */
  id: string;
  title: string;
  /** True only after the live GitHub API confirmed this step */
  verified: boolean;
  /**
     * ISO 8601 timestamp of first verified pass
     * @nullable
     */
  verifiedAt: string | null;
}

export interface DrillCandidate {
  /** Drill item id from the client-side drill bank */
  id: string;
  /**
     * Lesson/crisis id whose grader friction should boost this item's priority
     * @nullable
     */
  sourceId?: string | null;
}

export interface DrillCandidateSet {
  candidates: DrillCandidate[];
}

export interface DrillItemStats {
  id: string;
  /** How many times this item has been attempted */
  seenCount: number;
  correctCount: number;
  /**
     * Whether the most recent attempt was correct (null if never attempted)
     * @nullable
     */
  lastCorrect: boolean | null;
  /**
     * ISO 8601 timestamp of the most recent attempt
     * @nullable
     */
  lastSeenAt: string | null;
  /**
     * When this item comes due again (null means due now / never attempted)
     * @nullable
     */
  dueAt: string | null;
  dueNow: boolean;
  /** Scheduling priority; grader friction on the item's source raises it */
  priority: number;
}

export interface DrillDueResult {
  /** All candidates with stats, due items first in priority order */
  items: DrillItemStats[];
  dueCount: number;
}

export interface DrillAttemptInput {
  itemId: string;
  correct: boolean;
  /** @nullable */
  sourceId?: string | null;
}

export interface CapstoneStatus {
  githubConnected: boolean;
  /**
     * The connected GitHub username, when available
     * @nullable
     */
  githubLogin: string | null;
  repo: CapstoneRepo | null;
  /** @nullable */
  prNumber: number | null;
  /** @nullable */
  prUrl: string | null;
  /**
     * Head branch of the Dojo-generated pull request
     * @nullable
     */
  prBranch: string | null;
  missions: CapstoneMission[];
  /**
     * Set only when every mission has been verified against GitHub
     * @nullable
     */
  badgeEarnedAt: string | null;
}

export interface CapstoneVerifyResult {
  missionId: string;
  verified: boolean;
  /** Honest explanation of what the live check saw */
  detail: string;
  status: CapstoneStatus;
}

export interface CheckResult {
  ran: boolean;
  /**
     * True when every check passed, null when it could not be determined
     * @nullable
     */
  passed: boolean | null;
  /** Raw output of check.sh */
  output: string;
}

export interface CrisisScenario {
  /** Scenario id such as crisis-01 */
  id: string;
  number: number;
  /** The scenario's practice repo has been set up */
  hasPlayground: boolean;
  /** The playground contains a git repository */
  initialized: boolean;
  /** The grader has genuinely passed at least once */
  solved: boolean;
  /** Absolute path of the scenario's practice repository */
  path: string;
}

export interface CrisisSetupResult {
  ok: boolean;
  message: string;
  /** Absolute path of the scenario's practice repository */
  path: string;
}