# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /crisis/scenarios/{crisisId}/setup:
    post:
      operationId: setupCrisisScenario
      tags: [crisis]
      summary: Initialize the scenario's broken repository
      description: Idempotently rebuilds the scenario's practice repo in its broken state
      parameters:
        - name: crisisId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Setup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrisisSetupResult"
        "404":
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /crisis/scenarios/{crisisId}/state:
    get:
      operationId: getCrisisRepoState
      tags: [crisis]
      summary: Repository state for a crisis scenario playground
      description: Full git state of the scenario's practice repository
      parameters:
        - name: crisisId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Repository state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepoState"
        "404":
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /crisis/scenarios/{crisisId}/check:
    post:
      operationId: runCrisisCheck
      tags: [crisis]
      summary: Grade the scenario's recovery
      description: Runs the scenario's server-side grader against the practice repo
      parameters:
        - name: crisisId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Check results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckResult"
        "404":
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RepoFile:
      type: object
      properties:
        path:
          type: string
        status:
          type: string
          enum:
            [staged, modified, staged_and_modified, untracked, deleted, conflicted]
      required: [path, status]
    RepoCommit:
      type: object
      properties:
        hash:
          type: string
        shortHash:
          type: string
        subject:
          type: string
        authorName:
          type: string
        date:
          type: string
          description: ISO 8601 author date
        refs:
          type: array
          items:
            type: string
          description: Branch or tag names pointing at this commit
        branch:
          type: ["string", "null"]
          description: Branch this commit was listed under, when walking per-branch history
      required: [hash, shortHash, subject, authorName, date, refs]
    RepoBranch:
      type: object
      properties:
        name:
          type: string
        isCurrent:
          type: boolean
        headHash:
          type: string
      required: [name, isCurrent, headHash]
    RepoState:
      type: object
      properties:
        lessonId:
          type: string
        hasPlayground:
          type: boolean
        initialized:
          type: boolean
        currentBranch:
          type: ["string", "null"]
        detachedHead:
          type: boolean
        mergeInProgress:
          type: boolean
        files:
          type: array
          items:
            $ref: "#/components/schemas/RepoFile"
        commits:
          type: array
          items:
            $ref: "#/components/schemas/RepoCommit"
        branches:
          type: array
          items:
            $ref: "#/components/schemas/RepoBranch"
        remotes:
          type: array
          items:
            type: string
        summary:
          type: string
          description: Plain-English explanation of the current repository state
        hasBot:
          type: boolean
          description: The lesson has a scripted teammate (bot.sh) that can act when time passes
      required:
        [
          lessonId,
          hasBot,
          hasPlayground,
          initialized,
          currentBranch,
          detachedHead,
          mergeInProgress,
          files,
          commits,
          branches,
          remotes,
          summary,
        ]
...
    CheckResult:
      type: object
      properties:
        ran:
          type: boolean
        passed:
          type: ["boolean", "null"]
          description: True when every check passed, null when it could not be determined
        output:
          type: string
          description: Raw output of check.sh
      required: [ran, passed, output]
...
    CrisisSetupResult:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        path:
          type: string
          description: Absolute path of the scenario's practice repository
      required: [ok, message, path]